fixed test, improved file upload

+++ metadata-service/src/main/java/sanity/nil/meta/dto/user/SubscriptionDTO.java
@@ -1,11 +1,9 @@
 package sanity.nil.meta.dto.user;
 
-import java.math.BigDecimal;
-
 public record SubscriptionDTO(
         Short id,
         String title,
         String description,
-        BigDecimal storageLimit,
+        Long storageLimit,
         Integer workspacesLimit
 ) { }
